+++ test/integration/tests/3315-multi-ghc-options/Main.hs
@@ -2,8 +2,5 @@ import StackTest
 
 main :: IO ()
 main = do
-  return ()
-  {- FIXME: Temporarily disabled.  See #3353 / #3315
-  stack ["build", "--ghc-options=-DBAR -DBAZ"]
-  stack ["setup", "--ghcjs-boot-options=-DBAR -DBAZ"]
-  -}
+  stack ["build", "--ghc-options=-ddump-simpl -ddump-asm -DBAR -DBAZ"]
+  repl ["--ghc-options=-ddump-simpl -ddump-asm"] (pure ())