+++ ChangeLog.md
@@ -60,6 +60,12 @@ Major changes:
       is uniquely identified by a commit id and an Hadrian "flavour" (Hadrian is
       the newer GHC build system), hence `compiler` can be set to use a GHC
       built from source with `ghc-git-COMMIT-FLAVOUR`
+* `stack.yaml` now supports a `configure-options`, which are passed directly to
+  the `configure` step in the Cabal build process. See
+  [#1438](https://github.com/commercialhaskell/stack/issues/1438)
+
+* Remove support for building GHCJS itself. Future releases of Stack
+  may remove GHCJS support entirely.
 * Support for lock files for pinning exact project dependency versions
 
 Behavior changes:
@@ -113,6 +119,12 @@ Behavior changes:
 * For GHC 8.4 and later, disable the "shadowed dependencies" workaround. This
   means that Stack will no longer have to force reconfigures as often. See
   [#3554](https://github.com/commercialhaskell/stack/issues/3554).
+
+* Stack will check occassionally if there is a new version available and prompt
+  the user to upgrade. This will not incur any additional network traffic, as
+  it will piggy-back on the existing Hackage index updates. You can set
+  `recommend-stack-upgrade: false` to bypass this. See
+  [#1681](https://github.com/commercialhaskell/stack/issues/1681).
 
 Other enhancements:
 
@@ -216,6 +228,8 @@ Other enhancements:
   new-build`. See
   [#4706](https://github.com/commercialhaskell/stack/issues/4706).
 * Use a database cache table to speed up discovery of installed GHCs
+* You can specify multiple `--test-arguments` options. See
+  [#2226](https://github.com/commercialhaskell/stack/issues/2226)
 
 Bug fixes:
 
@@ -271,6 +285,8 @@ Bug fixes:
 - Docker: fix detection of expected subprocess failures.  This fixes
   downloading a compatible `stack` executable  when the host `stack` is not compatible with the Docker image (on Linux), and doesn't show an unnecessary
   extra error when the in-container re-exec'ed `stack` exits with failure.
+* The `stack ghci` command's `--ghc-options` flag now parses multiple options.
+  See [#3315](https://github.com/commercialhaskell/stack/issues/3315).
 
 ## v1.9.3